test: fix the test cases error.
