+++ tests/system/Router/Controllers/App/Dash_folder/Dash_controller.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of CodeIgniter 4 framework.
+ *
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
+namespace App\Controllers\Dash_folder;
+
+use CodeIgniter\Controller;
+
+class Dash_controller extends Controller
+{
+    public function getSomemethod(): void
+    {
+    }
+
+    public function getDash_method(): void
+    {
+    }
+}
