+++ tests/system/Router/Controllers/App/Mycontroller.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of CodeIgniter 4 framework.
+ *
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
+namespace App\Controllers;
+
+use CodeIgniter\Controller;
+
+class Mycontroller extends Controller
+{
+    public function getIndex(): void
+    {
+    }
+
+    public function getSomemethod($first = ''): void
+    {
+    }
+}
