+++ tests/system/Router/Controllers/App/Φ.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * This file is part of CodeIgniter 4 framework.
+ *
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
+namespace App\Controllers;
+
+use CodeIgniter\Controller;
+
+class Φ extends Controller
+{
+    public function getIndex(): void
+    {
+    }
+}
